requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
python-dotenv==1.0.1

# Streamlit UI
streamlit>=1.32.0

# AWS
boto3==1.35.36
botocore==1.35.36

# LangGraph Agent
langgraph==0.2.28
langchain==0.3.7
langchain-aws==0.2.7
langchain-core==0.3.15

# Overture Maps — Real DuckDB Parquet
duckdb==1.1.3
pyarrow==17.0.0

# Geospatial
shapely==2.0.6
folium==0.18.0

# Data
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2

# HTTP
httpx==0.27.2

# PDF generation
reportlab==4.2.2
